requirements.txt
openai
python-docx
docx2pdf
python-dotenv
flask
